refactor: Move `StorageService` to wallet library (#30922)
<!--
Please submit this PR as a draft initially.

Do not mark it as "Ready for review" until this PR meets the canonical
Definition of Ready For Review in `docs/readme/ready-for-review.md`.

In short: the template must be materially complete (not just section
titles
present), all status checks must be currently passing, and the only
expected
follow-up commits must be reviewer-driven.
-->

## **Description**

Move `StorageService` initialization to the wallet library. Since the
storage adapter was placed in the init file this PR also moves that
along with its test to somewhere else.

Also bumps `keyring-controller` to match the latest version shipped with
`wallet`.

No functional changes.

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry: null

## **Related issues**

https://consensyssoftware.atlassian.net/browse/WPC-1027

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Medium Risk**
> Touches persistent controller storage wiring and upgrades
wallet/keyring-controller; behavior should be equivalent but regressions
could affect on-disk data or keyring flows if wallet init differs
subtly.
> 
> **Overview**
> **`StorageService` is no longer bootstrapped as its own Engine
messenger client.** The dedicated init module, messenger factory, and
their tests are removed, and `Engine` no longer registers
`StorageService` in `initMessengerClients`.
> 
> **Persistence is wired through `@metamask/wallet` v2 instead.**
`initializeWallet` passes `instanceOptions.storageService.storage` using
the same **`mobileStorageAdapter`** logic, which is **moved** from
`storage-service-init` into `storage-service-utils` (adapter
implementation + tests consolidated there). Key encoding behavior is
unchanged.
> 
> **Dependencies:** `@metamask/wallet` is bumped to **^2.0.0** (which
brings in `@metamask/storage-service` and
**`@metamask/keyring-controller` ^26.0.0**), with mobile’s direct
`keyring-controller` dependency aligned to **^26.0.0**.
> 
> Intended as a structural refactor with **no functional change** to how
filesystem-backed controller storage works.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
35dd5b340d8499b7187cbf3e6991f9c79fa991e7. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->
