+++ app/core/Engine/wallet-init/initialization.ts
@@ -3,6 +3,7 @@ import { Json } from '@metamask/utils';
 import { getKeyringBuilders } from './keyrings';
 import { RootMessenger } from '../types';
 import { Encryptor, LEGACY_DERIVATION_OPTIONS } from '../../Encryptor';
+import { mobileStorageAdapter } from '../utils/storage-service-utils';
 
 export function initializeWallet({
   messenger,
@@ -22,6 +23,9 @@ export function initializeWallet({
       keyringController: {
         encryptor,
         keyringBuilders: getKeyringBuilders(messenger),
+      },
+      storageService: {
+        storage: mobileStorageAdapter,
       },
     },
   });
